requirements.txt
dotenv==0.9.9
faster-whisper==1.1.1
pillow==11.1.0
PyAudio==0.2.14
PyGetWindow==0.0.9
ttkbootstrap==1.10.1
keyboard==0.13.5
mem0ai==0.1.79
vecs==0.4.5
soundfile==0.13.1
pywin32
google-genai
mecab==0.996.11
chromadb==0.5.4
SpeechRecognition
kokoro
playwright
google.generativeai